new_agno_agent/requirements.txt
agno-agent>=1.0.0
streamlit>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pandas>=1.2.0
fpdf2>=2.5.0
python-dotenv>=0.15.0
groq>=0.1.0